Migrate educationService into new app directory - Refactor: opt for functional database interface over to database collection object - Refactor: split education and certification database interfaces
